[blog] iterators - move FixedArray class to /src
